Back-end completo

+++ src/utils/cloudinary/deleteContentCloudinary.ts
@@ -1,11 +1,22 @@
 import { v2 as cloudinary } from "cloudinary";
-import { FileProps } from "../../entities/DigitalContentEntity.js";
+import { FileProps } from "../../interfaces/FilePropsInterface";
 
 export const deleteContentCloudinary = async (images: FileProps[]) => {
-  const paths: string[] = [];
-  for (let img of images) {
-    paths.push(img.publicId);
-  }
+  try {
+    if (images.length > 0) {
+      const paths: string[] = [];
+      for (let img of images) {
+        paths.push(img.publicId);
+      }
 
-  await cloudinary.api.delete_resources(paths);
+      // <--- Delete images --->
+      await cloudinary.api.delete_resources(paths);
+
+      // <--- Delete videos --->
+      await cloudinary.api.delete_resources(paths, { resource_type: "video" });
+    }
+  } catch (error) {
+    console.log("Error ao deletar o arquivo");
+    console.log(error);
+  }
 };
